rename lobarocoap subdirectory (option-types) to fix build issues for ti.cc3220_launchpad
