+++ libraries/3rdparty/CoAp/coap.h
@@ -54,11 +54,11 @@
 
     #include "coap_options.h"
     #include "coap_message.h"
-    #include "option-types/coap_option_blockwise.h"
-    #include "option-types/coap_option_ETag.h"
-    #include "option-types/coap_option_cf.h"
-    #include "option-types/coap_option_uri.h"
-    #include "option-types/coap_option_observe.h"
+    #include "option_types/coap_option_blockwise.h"
+    #include "option_types/coap_option_ETag.h"
+    #include "option_types/coap_option_cf.h"
+    #include "option_types/coap_option_uri.h"
+    #include "option_types/coap_option_observe.h"
     #include "coap_resource.h"
     #include "coap_interaction.h"
     #include "coap_main.h"
