+++ libraries/3rdparty/CoAp/option_types/coap_option_uri.h
